% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colours.R
\docType{data}
\name{aiti_black}
\alias{aiti_black}
\title{Hex code for the colour: AITI Black}
\format{
An object of class \code{character} of length 1.
}
\usage{
aiti_black
}
\description{
#000000
}
\keyword{datasets}
