% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colours.R
\docType{data}
\name{aiti_blue_2}
\alias{aiti_blue_2}
\title{Hex code for the colour: AITI Blue 2 (Blues set)}
\format{
An object of class \code{character} of length 1.
}
\usage{
aiti_blue_2
}
\description{
#0044AA
}
\keyword{datasets}
