% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colours.R
\docType{data}
\name{aiti_darkblue}
\alias{aiti_darkblue}
\title{Hex code for the colour: AITI Dark Blue}
\format{
An object of class \code{character} of length 1.
}
\usage{
aiti_darkblue
}
\description{
#22233C
}
\keyword{datasets}
