% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colours.R
\docType{data}
\name{aiti_green}
\alias{aiti_green}
\title{Hex code for the colour: AITI Green}
\format{
An object of class \code{character} of length 1.
}
\usage{
aiti_green
}
\description{
#5aa2ae
}
\keyword{datasets}
