% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colours.R
\docType{data}
\name{aiti_grey}
\alias{aiti_grey}
\title{Hex code for the colour: AITI Grey}
\format{
An object of class \code{character} of length 1.
}
\usage{
aiti_grey
}
\description{
#b2b2b2
}
\keyword{datasets}
