% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colours.R
\docType{data}
\name{aiti_greyblue}
\alias{aiti_greyblue}
\title{Hex code for the colour: AITI Grey Blue}
\format{
An object of class \code{character} of length 1.
}
\usage{
aiti_greyblue
}
\description{
#1a5d9b
}
\keyword{datasets}
