% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colours.R
\docType{data}
\name{aiti_highlight_blue}
\alias{aiti_highlight_blue}
\title{Hex code for the colour: AITI Highlight Blue}
\format{
An object of class \code{character} of length 1.
}
\usage{
aiti_highlight_blue
}
\description{
#0077ff
}
\keyword{datasets}
