% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colours.R
\docType{data}
\name{aiti_highlight_yellow}
\alias{aiti_highlight_yellow}
\title{Hex code for the colour: AITI Highlight Yellow}
\format{
An object of class \code{character} of length 1.
}
\usage{
aiti_highlight_yellow
}
\description{
#FCAF2B
}
\keyword{datasets}
