% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colours.R
\docType{data}
\name{aiti_lightblue}
\alias{aiti_lightblue}
\title{Hex code for the colour: AITI Light Blue}
\format{
An object of class \code{character} of length 1.
}
\usage{
aiti_lightblue
}
\description{
#accbf9
}
\keyword{datasets}
