% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colours.R
\docType{data}
\name{aiti_palette_set}
\alias{aiti_palette_set}
\title{A list of aiti colour sets}
\format{
An object of class \code{list} of length 1.
}
\usage{
aiti_palette_set
}
\description{
"full": darkblue
}
\keyword{datasets}
