% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colours.R
\docType{data}
\name{aiti_purple}
\alias{aiti_purple}
\title{Hex code for the colour: AITI Purple}
\format{
An object of class \code{character} of length 1.
}
\usage{
aiti_purple
}
\description{
#9999ff
}
\keyword{datasets}
