% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aiti_scales.R
\name{aiti_scale}
\title{Title}
\usage{
aiti_colour_manual(
  n = 0,
  reverse = FALSE,
  discrete = TRUE,
  faded = FALSE,
  palette = "full",
  ...
)
}
\arguments{
\item{...}{}
}
\value{

}
\description{
Title
}
\examples{

}
