% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colours.R
\docType{data}
\name{aiti_yellow}
\alias{aiti_yellow}
\title{Hex code for the colour: AITI Yellow}
\format{
An object of class \code{character} of length 1.
}
\usage{
aiti_yellow
}
\description{
#fdd969
}
\keyword{datasets}
