% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aiti_scales.R
\name{scale_colour_aiti}
\alias{scale_colour_aiti}
\title{Colour scale with AITI palettes.}
\usage{
scale_colour_aiti(palette = "main", discrete = TRUE, reverse = FALSE, ...)
}
\arguments{
\item{palette}{}

\item{discrete}{}

\item{reverse}{}

\item{...}{}
}
\description{
Colour scale with AITI palettes.
}
