% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aiti_scales.R
\name{scale_fill_aiti}
\alias{scale_fill_aiti}
\title{Fill scale with AITI palettes.}
\usage{
scale_fill_aiti(palette = "main", discrete = TRUE, reverse = FALSE, ...)
}
\arguments{
\item{palette}{}

\item{discrete}{}

\item{reverse}{}

\item{...}{}
}
\value{

}
\description{
Fill scale with AITI palettes.
}
