% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_aiti.R
\name{theme_aiti}
\alias{theme_aiti}
\title{AITI Theme}
\usage{
theme_aiti(
  base_size = 12,
  colour = "Soft Black",
  base_family = "sans",
  markdown = FALSE,
  flipped = FALSE
)
}
\arguments{
\item{base_size}{The base size of text elements of the plot.}

\item{colour}{The background colour of the plot.}

\item{base_family}{The base family of text elements of the plot. The AITI default font is Space Mono but this is not
available on all systems (yet?).}

\item{markdown}{Include \code{ggtext::element_markdown}}

\item{flipped}{TRUE to flip the y-axis guide lines to show on the x-axis instead.}
}
\value{
ggplot2 theme
}
\description{
Theme for plots in AITI publications, reports, Shiny Apps.
}
