% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aiti_scales.R
\name{aiti_scale}
\alias{aiti_scale}
\alias{aiti_colour_manual}
\title{Set manual colours}
\usage{
aiti_colour_manual(
  n = 0,
  reverse = FALSE,
  discrete = TRUE,
  faded = FALSE,
  palette = "full",
  ...
)
}
\arguments{
\item{n}{number of colours}

\item{reverse}{(logical) whether to reverse the order of the colours}

\item{discrete}{(logical)}

\item{faded}{(logical) NYI}

\item{palette}{NYI}

\item{...}{other elements passed to ggplot2::scales}
}
\description{
Set manual colours
}
