% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_aiti.R
\name{theme_aiti}
\alias{theme_aiti}
\title{Title}
\usage{
theme_aiti(
  base_size = 13,
  base_family = "sans",
  chart_type = "normal",
  flipped = FALSE,
  background = "white",
  legend = "none",
  panel_borders = FALSE
)
}
\value{

}
\description{
Title
}
