% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_aiti.R
\name{theme_aiti}
\alias{theme_aiti}
\title{AITI Theme}
\usage{
theme_aiti(
  base_size = 12,
  colour = "blue",
  base_family = "sans",
  title_family = "mono",
  markdown = FALSE,
  flipped = FALSE
)
}
\arguments{
\item{base_size}{}

\item{base_family}{}

\item{title_family}{Plot title font family}

\item{markdown}{Include \code{ggtext::element_markdown}}

\item{flipped}{}

\item{color}{The background colour of the plot. One of \code{'blue', 'yellow', 'grey'}.}
}
\value{

}
\description{
Theme for plots in AITI publications, reports, Shiny Apps.
}
\details{
This theme should be used with \code{\link{scale_colour_aiti()}}
}
\examples{

}
