% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_aiti.R
\name{theme_aiti}
\alias{theme_aiti}
\title{Theme a ggplot2 object - AITI style}
\usage{
theme_aiti(
  base_size = 13,
  base_family = "sans",
  chart_type = "normal",
  flipped = FALSE,
  background = "white",
  legend = "none",
  panel_borders = FALSE
)
}
\arguments{
\item{base_size}{font size}

\item{base_family}{font family}

\item{chart_type}{scatter or normal (default = normal)}

\item{flipped}{NYI}

\item{background}{NYI}

\item{legend}{legend position}

\item{panel_borders}{(logical) default FALSE}
}
\description{
Theme a ggplot2 object - AITI style
}
